cambios

+++ proyecto/src/components/ItemTable.jsx
@@ -7,44 +7,71 @@ import {
 } from "reactstrap";
 import { useState } from "react";
 import "../styles/TableSalas.css";
+import Swal from "sweetalert2";
 
-const ItemTable = ({ sala, handleEdit, getData }) => {
+const ItemTable = ({ sala, handleEdit, getData, handleReserva }) => {
   const [dropdownOpen, setDropdownOpen] = useState(false);
 
   const toggle = () => setDropdownOpen((prevState) => !prevState);
 
   const onDelete = async () => {
-    const response = await fetch(`http://127.0.0.1:8000/api/room/${sala.id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
+    Swal.fire({
+      title: "¿Estás seguro?",
+      text: "Una vez eliminada, no podrás recuperar la sala",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí, eliminar",
+      cancelButtonText: "Cancelar",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          const response = await fetch(
+            `http://localhost:3001/salas/${sala.id}`,
+            {
+              method: "DELETE",
+            }
+          );
+          if (response.ok) {
+            getData();
+            Swal.fire("¡Sala eliminada!", "", "success");
+          }
+        } catch (error) {
+          Swal.fire("¡Error!", "No se pudo eliminar la sala", "error");
+        }
+      }
     });
-    if (response.ok) {
-      getData();
-      toggle();
-    } else {
-      console.log("Error al eliminar la sala");
-    }
   };
 
   return (
     <tr key={sala.id}>
       <th scope="row">{sala.name}</th>
       <td>{sala.capacity}</td>
-      <td>{sala.available ? "Disponible" : "No disponible"}</td>
-      <td>
+      <td className="item_table_salas_eventos">
         Esta sala tiene {sala.reservations} reservas para hoy{" "}
-        <Button color="danger" id="boton-reservar">
+        <Button
+          color="danger"
+          id="boton-reservar"
+          onClick={() => handleReserva(sala)}
+        >
           Reservar
         </Button>
         <Dropdown isOpen={dropdownOpen} toggle={toggle}>
-          <DropdownToggle data-toggle="dropdown" tag="span" id="puntos">
+          <DropdownToggle
+            data-toggle="dropdown"
+            tag="span"
+            style={{ cursor: "pointer" }}
+          >
             ...
           </DropdownToggle>
-          <DropdownMenu id="menuPuntos">
-            <DropdownItem onClick={() => handleEdit(sala)}>Editar</DropdownItem>
-            <DropdownItem onClick={onDelete}>Eliminar</DropdownItem>
+          <DropdownMenu>
+            <DropdownItem text onClick={() => handleEdit(sala)}>
+              Editar
+            </DropdownItem>
+            <DropdownItem text onClick={onDelete}>
+              Eliminar
+            </DropdownItem>
           </DropdownMenu>
         </Dropdown>
       </td>
